+++ typed/typed_test.go
@@ -1,4 +1,4 @@
-package typedData
+package typed
 
 import (
 	"encoding/json"
